ref(performance): Move PerformanceSearchQueryBuilder into its own component
This component will be reused anywhere we want to search for transactions.
